# backend/openapi/deadmans.v1.yaml
openapi: 3.0.3
info:
  title: Dead-Mans API
  version: v1
  description: >
    Canonical transport contract for Dead-Mans: HTTP paths under /api and SignalR hubs
    documented in x-signalr. Regenerate frontend contracts after changes.
servers:
  - url: /api
paths:
...
  /game/questions/twitch-preview:
    post:
      operationId: previewTwitchQuizMessages
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwitchQuizPreviewRequestDto'
      responses:
        '200':
          description: Normalized Twitch message preview and length validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwitchQuizPreviewDto'
        '400':
          description: Invalid preview input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /game/questions/{questionId}/enabled:
    patch:
      operationId: setGameQuestionEnabled
      security:
        - cookieAuth: []
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetGameQuestionEnabledRequestDto'
      responses:
        '204':
          description: Question enable flag updated
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Question or question category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /game/questions/{questionId}:
    put:
      operationId: updateGameQuestion
      security:
        - cookieAuth: []
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGameQuestionRequestDto'
      responses:
        '200':
          description: Question updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameQuestionCatalogItemDto'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Question not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteGameQuestion
      security:
        - cookieAuth: []
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Question soft-deleted from catalog
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Question not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /game/history/games/{gameId}:
    get:
      operationId: getGameHistoryGameDetails
      security:
        - cookieAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detailed game history with separate main-game and quiz sections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameHistoryGameDetailsDto'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /game/lifecycle/games/{gameId}:
    delete:
      operationId: archiveGame
      security:
        - cookieAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Non-draft game archived via soft-delete
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Draft games must be deleted through game setup draft deletion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  responses:
...
    InternalServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dm_auth
  schemas:
...
    GameFinishSummaryDto:
      type: object
      required:
        - gameId
        - gameTitle
        - gameStatus
        - boardVersion
        - calculationVersion
        - completedRoundCount
        - cancelledRoundCount
        - totalKills
        - totalBounties
        - quizTotalPoints
        - pendingQuizQuestionCount
        - skippedQuizQuestionCount
        - teams
      properties:
        gameId:
          type: string
          format: uuid
        gameTitle:
          type: string
        gameStatus:
          type: string
          enum: [active, finished]
        boardVersion:
          type: integer
          minimum: 1
        finishedAtUtc:
          type: string
          format: date-time
          nullable: true
        finishedByUserId:
          type: string
          format: uuid
          nullable: true
        finishedByDisplayName:
          type: string
          nullable: true
        publicNote:
          type: string
          nullable: true
          maxLength: 2000
        calculationVersion:
          type: integer
          minimum: 1
        completedRoundCount:
          type: integer
        cancelledRoundCount:
          type: integer
        totalKills:
          type: integer
        totalBounties:
          type: integer
        quizTotalPoints:
          type: integer
        pendingQuizQuestionCount:
          type: integer
        skippedQuizQuestionCount:
          type: integer
        teams:
          type: array
          items:
            $ref: '#/components/schemas/GameFinishTeamResultDto'
...
    GameModifierActivationLimitDto:
      type: object
      properties:
        count:
          type: integer
          nullable: true
          minimum: 1
...
    GameQuestionOptionInputDto:
      type: object
      required: [text, isCorrect]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
        isCorrect:
          type: boolean
    GameQuestionOptionDto:
      type: object
      required: [optionId, text, isCorrect, sortOrder]
      properties:
        optionId:
          type: string
          format: uuid
        text:
          type: string
        isCorrect:
          type: boolean
        sortOrder:
          type: integer
          minimum: 0
    GameQuestionCatalogItemDto:
      type: object
      required:
        - questionId
        - questionCode
        - categoryId
        - categoryName
        - text
        - options
        - reward
        - priority
        - isEnabled
        - askedTotalCount
        - submissionTotalCount
        - correctSubmissionTotalCount
        - correctPercentage
        - twitchCompatible
      properties:
        questionId:
          type: string
          format: uuid
        questionCode:
          type: string
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
          description: Human-readable name of the question category.
        text:
          type: string
        options:
          type: array
          minItems: 2
          maxItems: 10
          items:
            $ref: '#/components/schemas/GameQuestionOptionDto'
        reward:
          type: integer
          minimum: 0
        priority:
          type: integer
        isEnabled:
          type: boolean
        askedTotalCount:
          type: integer
          minimum: 0
        submissionTotalCount:
          type: integer
          minimum: 0
        correctSubmissionTotalCount:
          type: integer
          minimum: 0
        correctPercentage:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
        lastAskedAtUtc:
          type: string
          format: date-time
          nullable: true
        twitchCompatible:
          type: boolean
        twitchCompatibilityErrorCode:
          type: string
          nullable: true
    TwitchQuizPreviewRequestDto:
      type: object
      required: [text, options, durationSeconds, reward]
      properties:
        text: { type: string, nullable: true }
        options:
          type: array
          minItems: 2
          maxItems: 10
          items: { type: string, nullable: true }
        durationSeconds: { type: integer, minimum: 5, maximum: 3600 }
        reward: { type: integer, minimum: 0 }
    TwitchQuizPreviewDto:
      type: object
      required: [question, options, resultTemplate, questionLength, optionsLength, resultMaximumLength, isCompatible, maximumLength]
      properties:
        question: { type: string }
        options: { type: string }
        resultTemplate: { type: string }
        questionLength: { type: integer }
        optionsLength: { type: integer }
        resultMaximumLength: { type: integer }
        isCompatible: { type: boolean }
        errorCode: { type: string, nullable: true }
        maximumLength: { type: integer, enum: [500] }
...
    SetGameQuestionEnabledRequestDto:
      type: object
      required: [isEnabled]
      properties:
        isEnabled:
          type: boolean
...
    UpdateGameQuestionRequestDto:
      type: object
      description: Replaces the complete option set; published games retain their snapshots.
      required: [categoryId, text, options, reward]
      properties:
        categoryId:
          type: string
          format: uuid
        text:
          type: string
          minLength: 1
          maxLength: 2000
        options:
          type: array
          minItems: 2
          maxItems: 10
          items:
            $ref: '#/components/schemas/GameQuestionOptionInputDto'
        reward:
          type: integer
          minimum: 0
        isEnabled:
          type: boolean
          default: true
        priority:
          type: integer
          default: 0
...
    GameHistoryPlayerSummaryDto:
      type: object
      required:
        - userId
        - displayName
        - points
        - eventCount
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
        points:
          type: integer
        eventCount:
          type: integer
          minimum: 0
        lastActivityAtUtc:
          type: string
          format: date-time
          nullable: true
    GameHistoryQuizPlayerSummaryDto:
      type: object
      required: [userId, displayName, points, attempts, correctAnswers]
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
        points:
          type: integer
        attempts:
          type: integer
          minimum: 0
        correctAnswers:
          type: integer
          minimum: 0
        lastActivityAtUtc:
          type: string
          format: date-time
          nullable: true
    GameHistoryModifierActivationItemDto:
      type: object
      required:
        - activationId
        - modifierId
        - modifierName
        - activatedByUserId
        - activatedByDisplayName
        - activatedAtUtc
        - status
        - refundAmount
      properties:
        activationId:
          type: string
          format: uuid
        modifierId:
          type: string
          format: uuid
        modifierName:
          type: string
        activatedByUserId:
          type: string
          format: uuid
        activatedByDisplayName:
          type: string
        activatedAtUtc:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, consumed, cancelled]
        cancelledAtUtc:
          type: string
          format: date-time
          nullable: true
        refundAmount:
          type: integer
          minimum: 0
...
    GameHistoryTeamLeaderboardEntryDto:
      type: object
      required:
        - teamId
        - teamSlotIndex
        - roundsPlayed
        - bestScore
        - penaltyTotal
        - finalScore
        - bestRound
        - latestRound
        - rounds
        - totalScore
        - averageScore
        - totalBonusDelta
        - totalKills
        - totalBounties
        - participantNames
        - lastFinishedAtUtc
      properties:
        teamId:
          type: string
          format: uuid
        teamName:
          type: string
          nullable: true
        teamSlotIndex:
          type: integer
        roundsPlayed:
          type: integer
        bestScore:
          type: integer
        penaltyTotal:
          type: integer
        finalScore:
          type: integer
        bestRound:
          $ref: '#/components/schemas/GameHistoryRoundItemDto'
        latestRound:
          $ref: '#/components/schemas/GameHistoryRoundItemDto'
        rounds:
          type: array
          items:
            $ref: '#/components/schemas/GameHistoryRoundItemDto'
        totalScore:
          type: integer
        averageScore:
          type: integer
        totalBonusDelta:
          type: integer
        totalKills:
          type: integer
        totalBounties:
          type: integer
        participantNames:
          type: array
          items:
            type: string
        lastFinishedAtUtc:
          type: string
          format: date-time
    GameHistoryMainGameSectionDto:
      type: object
      required:
        - playerStats
        - teamStats
        - modifierActivations
        - rounds
      properties:
        playerStats:
          type: array
          items:
            $ref: '#/components/schemas/GameHistoryPlayerSummaryDto'
        teamStats:
          type: array
          items:
            $ref: '#/components/schemas/GameHistoryTeamLeaderboardEntryDto'
        modifierActivations:
          type: array
          items:
            $ref: '#/components/schemas/GameHistoryModifierActivationItemDto'
        rounds:
          type: array
          items:
            $ref: '#/components/schemas/GameHistoryRoundItemDto'
    GameHistoryQuizSectionDto:
      type: object
      required:
        - totalPoints
        - playerStats
        - questionSessions
        - manualAwards
      properties:
        totalPoints:
          type: integer
        playerStats:
          type: array
          items:
            $ref: '#/components/schemas/GameHistoryQuizPlayerSummaryDto'
        questionSessions:
          type: array
          items:
            $ref: '#/components/schemas/GameHistoryQuizQuestionSessionItemDto'
        manualAwards:
          type: array
          items:
            $ref: '#/components/schemas/GameHistoryQuizManualAwardItemDto'
    GameHistoryModifierSnapshotDto:
      type: object
      required: [modifierId, versionId, revision, name, description, category, activationCost, activationLimit, normalizedTags, behaviorV2, conflicts, successfulActivationsCount, cancelledActivationsCount, resultsCount, isEmergencyDisabled]
      properties:
        modifierId: { type: string, format: uuid }
        versionId: { type: string, format: uuid }
        revision: { type: integer, minimum: 1 }
        name: { type: string }
        description: { type: string }
        category: { type: string, enum: [preparation, round, result] }
        iconEmoji: { type: string, nullable: true }
        activationCommand: { type: string, nullable: true }
        activationCost: { type: integer, minimum: 0 }
        activationLimit: { $ref: '#/components/schemas/GameModifierActivationLimitDto' }
        normalizedTags:
          type: array
          items: { type: string }
        behaviorV2: { $ref: '#/components/schemas/GameModifierBehaviorV2' }
        conflicts:
          type: array
          items: { $ref: '#/components/schemas/ModifierConflictSnapshotDto' }
        successfulActivationsCount: { type: integer, minimum: 0 }
        cancelledActivationsCount: { type: integer, minimum: 0 }
        resultsCount: { type: integer, minimum: 0 }
        isEmergencyDisabled: { type: boolean }
        emergencyDisabledAtUtc: { type: string, format: date-time, nullable: true }
    GameHistoryGameDetailsDto:
      type: object
      required:
        - gameId
        - gameTitle
        - gameStatus
        - createdAtUtc
        - mainGame
        - quiz
        - modifierSnapshotStatus
        - modifierSnapshots
      properties:
        gameId:
          type: string
          format: uuid
        gameTitle:
          type: string
        gameStatus:
          type: string
        createdAtUtc:
          type: string
          format: date-time
        startedAtUtc:
          type: string
          format: date-time
          nullable: true
        finishedAtUtc:
          type: string
          format: date-time
          nullable: true
        mainGame:
          $ref: '#/components/schemas/GameHistoryMainGameSectionDto'
        quiz:
          $ref: '#/components/schemas/GameHistoryQuizSectionDto'
        finalResult:
          type: object
          allOf:
            - $ref: '#/components/schemas/GameFinishSummaryDto'
          nullable: true
        modifierSnapshotStatus:
          type: string
          enum: [complete, legacy_unavailable]
        modifierSnapshots:
          type: array
          items: { $ref: '#/components/schemas/GameHistoryModifierSnapshotDto' }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
          nullable: true
          description: Stable machine-readable error code.
          enum:
            - auth.api_client_header_required
            - role_administration.invalid_request
            - role_administration.user_not_found
            - role_administration.permanent_superadmin_protected
            - game_board.cell_not_found
            - game_board.active_team_required
            - game_board.active_team_no_active_game
            - game_board.active_team_not_found
            - game_board.active_team_not_confirmed
            - game_board.active_team_already_played
            - game_board.active_team_has_no_active_members
            - game_board.active_team_round_in_progress
            - game_board.team_played_state_no_active_game
            - game_board.team_played_state_not_found
            - game_board.team_played_state_not_confirmed
            - game_board.team_played_state_round_in_progress
            - game_setup.no_draft
            - game_setup.draft_exists
            - game_setup.invalid_title
            - game_setup.invalid_save_request
            - game_setup.cell_not_found
            - game_setup.cell_media_not_found
            - game_setup.invalid_cell_media_upload
            - game_setup.stale_version
            - game_lifecycle.draft_not_found
            - game_lifecycle.current_already_exists
            - game_lifecycle.active_already_exists
            - game_lifecycle.game_not_ready
            - game_lifecycle.game_not_active
            - game_lifecycle.registration_slots_required
            - game_lifecycle.questions_unavailable
            - game_lifecycle.invalid_team_size_limits
            - game_lifecycle.no_confirmed_teams
            - game_lifecycle.unconfirmed_teams
            - game_lifecycle.pending_invitations
            - game_lifecycle.pending_disband_requests
            - game_lifecycle.invalid_confirmed_team_roster
            - game_lifecycle.operation_failed
            - game_lifecycle.draft_delete_not_allowed
            - game_lifecycle.archive_not_allowed
            - game_lifecycle.game_not_found
            - game_finish.round_in_progress
            - game_finish.stale_version
            - game_finish.warnings_not_acknowledged
            - game_finish.modifier_state_invalid
            - game_finish.invalid_request
            - game_common.unexpected_server_error
            - game_common.too_many_requests
            - game_registration.not_open
            - game_registration.no_slots
            - game_registration.already_on_team
            - game_registration.team_not_found
            - game_registration.team_not_joinable
            - game_registration.team_roster_locked
            - game_registration.not_team_member
            - game_registration.invitation_invalid
            - game_registration.slot_not_found
            - game_registration.slot_not_available
            - game_registration.user_not_found
            - game_registration.pending_invitation
            - game_registration.pending_outgoing_invitation
            - game_registration.team_invite_not_allowed
            - game_registration.team_active_in_game
            - game_registration.team_already_played
            - game_registration.invalid_team_name
            - game_registration.team_name_taken
            - game_registration.team_name_required
            - game_registration.team_not_full
            - game_registration.disband_request_not_owned
            - game_registration.operation_failed
            - game_modifier.game_not_active
            - game_modifier.not_enabled
            - game_modifier.emergency_disabled
            - game_modifier_content_locked
            - game_modifier_revision_stale
            - game_modifier_compatibility_locked
            - game_modifier_archived
            - game_modifier_version_binding_missing
            - game_modifier.conflict_active
            - game_modifier.limit_reached
            - game_modifier.ordering_closed
            - game_modifier.active_team_member
            - game_modifier.insufficient_quiz_points
            - game_modifier.player_not_found
            - game_modifier.activation_not_found
            - game_modifier.activation_cancel_forbidden
            - game_modifier.activation_cancel_invalid_state
            - game_modifier.activation_cancel_reason_required
            - game_modifier.user_not_resolved
            - game_modifier.invalid_request
            - game_modifier_not_found
            - game_round.no_active_game
            - game_round.cell_not_found
            - game_round.cell_not_open
            - game_round.team_not_found
            - game_round.team_not_confirmed
            - game_round.team_has_no_active_members
            - game_round.awaiting_modifiers_required
            - game_round.already_in_progress
            - game_round.invalid_request
            - game_round.not_found
            - game_round.not_in_progress
            - game_round.stale_version
            - game_round.modifier_result_not_found
            - modifier_resolution.duplicate_group
            - modifier_resolution.duplicate_result
            - modifier_resolution.result_set_mismatch
            - modifier_resolution.group_set_mismatch
            - modifier_resolution.group_missing
            - modifier_resolution.group_members_mismatch
            - modifier_resolution.violation_comment_required
            - modifier_resolution.automatic_input_forbidden
            - modifier_resolution.boolean_required
            - modifier_resolution.non_negative_count_required
            - modifier_resolution.unsupported
            - modifier_resolution.missing
            - modifier_calculation.failed
            - behavior.invalid
            - behavior.rule_incompatible
            - formula.unsupported
            - formula.incompatible
            - resolution.invalid
            - round_facts.invalid
            - activation.duplicate
            - resolution.rule_status_required
            - resolution.automatic_required
            - resolution.boolean_required
            - resolution.non_negative_count_required
            - resolution.count_exceeds_resolved_kills
            - resolution.count_exceeds_activation_limit
            - resolution.per_activation_required
            - game_question.invalid_request
            - game_question.duplicate_code
            - game_question.not_found
            - game_question.category_not_found
            - game_question.category_not_empty
            - game_question.category_protected
            - game_question.import_invalid_fields
            - game_question.import_duplicate_code_in_file
            - game_question.import_category_unresolved
            - game_question.import_duplicate_code_existing
            - game_quiz.no_active_game
            - game_quiz.no_available_questions
            - game_quiz.answer_player_not_found
            - game_quiz.question_session_not_found
            - game_quiz.question_session_closed
            - game_quiz.already_answered
            - game_quiz.option_not_found
            - game_quiz.manual_award_player_not_found
            - game_quiz.manual_award_invalid_points
            - game_quiz.manual_award_invalid_operation
            - game_quiz.manual_award_invalid_reason
            - game_quiz.manual_award_insufficient_points
            - game_quiz.manual_award_duplicate_request_conflict
            - twitch_quiz.disabled
            - twitch_quiz.not_connected
            - twitch_quiz.publication_in_progress
            - twitch_quiz.pending_outcome
            - twitch_quiz.incompatible_question
            - twitch_quiz.question_message_too_long
            - twitch_quiz.options_message_too_long
            - twitch_quiz.result_message_too_long
        requestId:
          type: string
          nullable: true
          description: Server request correlation identifier for diagnostics.
...